refactor(#699): move engine logic out of the Drawing result object (slice a)
Drawing had re-accreted three pieces of engine logic that belong in the
pipeline layers below it (#699 audit):

- The balloon band-assignment solver (_FlowEdge/_strip_capacity/
  _assign_balloon_bands, min-cost max-flow #516) moves to layout.py with
  the other solvers (ADR 0003/0009).
- The balloon render pass (add_balloons body, _place_band,
  _render_balloon) moves to a new annotations/balloons.py, threaded
  through PlacementContext — the orchestrator now calls sideways into
  the render layer instead of up into dwg.add_balloons by duck-typing.
  Drawing.add_balloons stays as the thin public verb. Drawing's
  duplicate _feature_of_hole_at / _hole_feature_index copy is deleted
  (PlacementContext.feature_of_hole_at was already the one
  implementation).
- _build_table moves to _core beside _table_metrics (the #700 sizing
  model); ADR 0005 §1 annotated with why the planned tables.py was
  never needed.

The #639 encapsulation ratchet stays at its empty allowlist — the moved
pass touches the drawing only through its public surface. Slice (a) of
 #699; slices b–d follow.
